move get notifications to NotificationButtonWithMenu
